location points api done with kotlin coroutines

+++ lib/src/main/java/com/wethejumpingspiders/isslocationapi/sightingsparser/SightingInfo.kt
@@ -13,7 +13,7 @@ import com.wethejumpingspiders.isslocationapi.locationpointparser.LocationPoint
 )
 data class SightingInfo(
     @PrimaryKey(autoGenerate = true)
-    val id:Long = 0,
+    val id:Int = 0,
     val date: String, val time: String, val duration: String, val maxElevation: String,
-    val approach: String, val departure: String, val locationPointId: Long
+    val approach: String, val departure: String, val locationPointId: Int
 )
